debug: I did bunch of updates.
first i unified the name of all column so it is easier for me in the future to understand what is happening.

I removed (not null) so I can avoid depuging and cleaning.

I change the value of varchar to text, or increased the size of varchar.

I debug minor mistakes.